tsARIBCharset:  Fix VS2015 build error
Added static_cast to fix build error that occurs with VS2015.

+++ src/libtsduck/dtv/charset/tsARIBCharset.cpp
@@ -325,7 +325,7 @@ bool ts::ARIBCharset::Decoder::decodeOneChar(const CharMap* gset)
 
         // Insert code point, if one was found.
         if (cp != 0) {
-            _str.append(cp);
+            _str.append(static_cast<uint32_t>(cp));
             return true;
         }
         else {
